Renamed src/functions/routers to interactivity
