+++ src/functions/interactivity/types.ts
