+++ src/functions/types.ts
@@ -15,7 +15,7 @@ import type SchemaTypes from "../schema/schema_types.ts";
 import type SlackSchemaTypes from "../schema/slack/schema_types.ts";
 import { SlackManifest } from "../manifest/mod.ts";
 
-export type { BlockActionHandler } from "./routers/types.ts";
+export type { BlockActionHandler } from "./interactivity/types.ts";
 
 export type FunctionInvocationBody = {
   "team_id": string;
