+++ src/mod.ts
@@ -6,7 +6,7 @@ export type {
 } from "./manifest/types.ts";
 export type { ManifestSchema } from "./manifest/manifest_schema.ts";
 export { DefineFunction } from "./functions/mod.ts";
-export { BlockActionsRouter } from "./functions/routers/mod.ts";
+export { BlockActionsRouter } from "./functions/interactivity/mod.ts";
 export { DefineWorkflow } from "./workflows/mod.ts";
 export { DefineType } from "./types/mod.ts";
 export { DefineOAuth2Provider } from "./providers/oauth2/mod.ts";
